+++ src/lib.rs
@@ -72,6 +72,7 @@
 pub mod bitreader;
 pub mod encoder;
 pub mod header;
+pub mod linepack;
 pub mod spec_tables;
 pub mod tables;